feat: 일정 제목 원시값을 객체로 포장
