+++ backend/src/main/java/com/allog/dallog/schedule/domain/Period.java
@@ -7,7 +7,7 @@ import javax.persistence.Embeddable;
 
 @Embeddable
 public class Period {
-
+    
     @Column(nullable = false)
     private LocalDateTime startDateTime;
