+++ backend/src/main/java/com/allog/dallog/schedule/domain/Schedule.java
@@ -14,15 +14,14 @@ import javax.persistence.Table;
 @Table(name = "SCHEDULES")
 public class Schedule {
 
-    private static final int MAX_TITLE_LENGTH = 20;
     private static final int MAX_MEMO_LENGTH = 255;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long Id;
 
-    @Column(length = 20, nullable = false)
-    private String title;
+    @Embedded
+    private Title title;
 
     @Embedded
     private Period period;
@@ -35,17 +34,10 @@ public class Schedule {
 
     public Schedule(final String title, final LocalDateTime startDateTime,
         final LocalDateTime endDateTime, final String memo) {
-        validateTitleLength(title);
         validateMemoLength(memo);
-        this.title = title;
+        this.title = new Title(title);
         this.period = new Period(startDateTime, endDateTime);
         this.memo = memo;
-    }
-
-    private void validateTitleLength(String title) {
-        if (title.length() > MAX_TITLE_LENGTH) {
-            throw new InvalidScheduleException("일정 제목의 길이는 20을 초과할 수 없습니다.");
-        }
     }
 
     private void validateMemoLength(String memo) {
@@ -59,7 +51,7 @@ public class Schedule {
     }
 
     public String getTitle() {
-        return title;
+        return title.getValue();
     }
 
     public LocalDateTime getStartDateTime() {
